An incorrect example of sendPayment with error handling
Judging by the code, the sendPayment function creates a promise where, in case of an error, it will reject. That is, an exception will be thrown with an error string but not as in the example - an error is returned as a response.

Moreover, in the code, an reject returns an object in some places, in others - a string.